Fixed errors and some warnings in MSVC build

See merge request GPUDB/nvcomp!320

+++ src/lowlevel/CascadedBatch.cu
@@ -495,7 +495,9 @@ __device__ void block_bitpack(
     // The bitpacking metadata consists of FOR (data_type size) and bitwidth and
     // the number of elements (4B). The start of the bitpacking data needs to be
     // both 4B and data_type aligned.
-    *out_bytes = roundUpTo(sizeof(data_type) + 4, max(4L, sizeof(data_type)))
+    *out_bytes = roundUpTo(
+                     sizeof(data_type) + 4,
+                     max(static_cast<size_t>(4), sizeof(data_type)))
                  + num_output_elements * sizeof(uint32_t);
   }
 
@@ -1275,7 +1277,7 @@ __device__ void cascaded_decompression_fcn(
           // aligned with 4B.
           rle_offsets[num_RLEs] = roundUpTo(
               rle_offsets[num_RLEs - 1] + chunk_metadata[num_RLEs],
-              max(4L, sizeof(data_type)));
+              max(static_cast<size_t>(4), sizeof(data_type)));
         }
       }
       __syncthreads();
